requirements.txt
torch==2.6.0
torchvision==0.21.0
git+https://github.com/huggingface/diffusers.git
transformers==4.49.0
accelerate==0.34.0
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.1.0
pydantic==2.10.0
huggingface_hub
numpy